requirements.txt
bs4==0.0.1
pymongo==3.7.1
python-decouple==3.4
python-dotenv==0.20.0
python-telegram-bot==12.7
requests==2.27.1
schedule==1.1.0